[S12P31B106-305] 엔티티 타입변경

See merge request s12-final/S12P31B106!265

+++ backend/src/main/java/com/voiceprint/backend/api/auth/dto/CustomOAuth2User.java
@@ -49,7 +49,7 @@ public class CustomOAuth2User implements OAuth2User {
     }
 
     // 사용자 ID 반환 (리프레시 토큰용)
-    public Long getUserId() {
+    public Integer getUserId() {
         return user.getId();
     }
 }
