+++ backend/src/main/java/com/voiceprint/backend/common/util/JWTUtil.java
@@ -62,7 +62,7 @@ public class JWTUtil {
      * @param userId 사용자 ID
      * @return 생성된 리프레시 토큰
      */
-    public String createRefreshToken(Long userId) {
+    public String createRefreshToken(Integer userId) {
         return Jwts.builder()
                 .setId(userId.toString()) // Redis 키로 사용할 ID 설정
                 .setSubject("refresh")
@@ -112,11 +112,11 @@ public class JWTUtil {
     /**
      * 토큰에서 사용자 ID를 추출합니다. (리프레시 토큰용)
      */
-    public Long getUserIdFromToken(String token) {
+    public Integer getUserIdFromToken(String token) {
         try {
             Claims claims = getAllClaims(token);
             String id = claims.getId();
-            return id != null ? Long.parseLong(id) : null;
+            return id != null ? Integer.parseInt(id) : null;
         } catch (Exception e) {
             return null;
         }
